Get the title and description tests passing; ditch the old transformWork
