+++ catalogue_pipeline/transformer/transformer_miro/src/test/scala/uk/ac/wellcome/platform/transformer/miro/generators/MiroRecordGenerators.scala
@@ -7,6 +7,7 @@ trait MiroRecordGenerators {
     title: Option[String] = Some("Auto-created title in MiroRecordGenerators"),
     creator: Option[List[Option[String]]] = None,
     description: Option[String] = None,
+    academicDescription: Option[String] = None,
     secondaryCreator: Option[List[String]] = None,
     artworkDate: Option[String] = None,
     copyrightCleared: Option[String] = Some("Y"),
@@ -29,6 +30,7 @@ trait MiroRecordGenerators {
       title = title,
       creator = creator,
       description = description,
+      academicDescription = academicDescription,
       secondaryCreator = secondaryCreator,
       artworkDate = artworkDate,
       copyrightCleared = copyrightCleared,
